Move validator_reg_message_factory.py to tests directory
Also fix up imports
