+++ consensus/poet/families/tests/validator_reg_message_factory.py
